Created Event Listener for a button
selected all the btns in the project by using
document.querySellectorAll("name of the class).
After that a loop is created to loop through the btns
using forEach and event listener was added to the
current target. from the target move to the parent
then to the grandParent to add a class called showDetails
using the classList.toggle method.

what this does is that anytime the plus button is click
it display the rest of the content which are hidden
by default.